# api/openapi/openapi3.json
{"components":{"requestBodies":{"CreateTodoRequest":{"content":{"application/json":{"schema":{"properties":{"is_done":{"type":"boolean"},"task":{"type":"string"}}}}},"description":"Request used for creating a todo.","required":true},"UpdateTodoRequest":{"content":{"application/json":{"schema":{"properties":{"is_done":{"type":"boolean"},"task":{"type":"string"}}}}},"description":"Request used for updating a todo.","required":true}},"responses":{"CreateTodoResponse":{"content":{"application/json":{"schema":{"properties":{"task":{"$ref":"#/components/schemas/Todo"}}}}},"description":"Response returned back after creating tasks."},"ErrorResponse":{"content":{"application/json":{"schema":{"properties":{"error":{"type":"string"}}}}},"description":"Response when errors happen."},"ReadTodoResponse":{"content":{"application/json":{"schema":{"properties":{"task":{"$ref":"#/components/schemas/Todo"}}}}},"description":"Response returned back after searching one task."}},"schemas":{"Todo":{"properties":{"id":{"format":"int64","type":"integer"},"is_done":{"type":"boolean"},"task":{"type":"string"}},"type":"object"}}},"info":{"contact":{"name":"Krishna","url":"https://github.com/KrisCatDog"},"description":"REST APIs example for standard boilerplate","license":{"name":"MIT","url":"https://opensource.org/licenses/MIT"},"title":"Go Standard Modular API HAHA","version":"1.0.0"},"openapi":"3.0.0","paths":{"/todos":{"get":{"operationId":"TodosList","responses":{"200":{"$ref":"#/components/responses/ReadTodoResponse"},"404":{"description":"Task not found"},"500":{"$ref":"#/components/responses/ErrorResponse"}},"tags":["todo"]},"post":{"operationId":"CreateTodo","requestBody":{"$ref":"#/components/requestBodies/CreateTodoRequest"},"responses":{"201":{"$ref":"#/components/responses/CreateTodoResponse"},"400":{"$ref":"#/components/responses/ErrorResponse"},"500":{"$ref":"#/components/responses/ErrorResponse"}},"tags":["todo"]}},"/todos/{id}":{"delete":{"operationId":"DeleteTodo","parameters":[{"in":"path","name":"id","required":true,"schema":{"format":"uuid","type":"string"}}],"responses":{"200":{"description":"Task updated"},"404":{"description":"Task not found"},"500":{"$ref":"#/components/responses/ErrorResponse"}},"tags":["todo"]},"get":{"operationId":"ReadTodo","parameters":[{"in":"path","name":"id","required":true,"schema":{"format":"uuid","type":"string"}}],"responses":{"200":{"$ref":"#/components/responses/ReadTodoResponse"},"404":{"description":"Task not found"},"500":{"$ref":"#/components/responses/ErrorResponse"}},"tags":["todo"]},"put":{"operationId":"UpdateTodo","parameters":[{"in":"path","name":"id","required":true,"schema":{"format":"uuid","type":"string"}}],"requestBody":{"$ref":"#/components/requestBodies/UpdateTodoRequest"},"responses":{"200":{"description":"Task updated"},"400":{"$ref":"#/components/responses/ErrorResponse"},"404":{"description":"Task not found"},"500":{"$ref":"#/components/responses/ErrorResponse"}},"tags":["todo"]}}},"servers":[{"description":"Local development","url":"http://localhost:1234"}],"tags":[{"name":"todo"}]}